requirements.txt
asgiref==3.7.2
class-registry==2.1.2
Django==4.2.1
django-filter==23.2
djangorestframework==3.14.0
psycopg-binary==3.1.9
psycopg2-binary==2.9.6
python-dateutil==2.8.2
pytz==2023.3
regex==2023.5.5
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.6.2
